some api reorganization
