+++ app/assets/javascripts/actions/magma_actions.js
@@ -1,5 +1,5 @@
 import { showMessages } from './message_actions';
-import { getDocuments, postRevisions } from '../api/timur';
+import { getTSVForm, getDocuments, postRevisions } from '../api/magma';
 import { Exchange } from './exchange_actions';
 
 export const addTemplate = (template)=>(
@@ -187,52 +187,9 @@ export const sendRevisions = (model_name, revisions, success, error)=>{
   }
 }
 
-/*
-export const sendRevisions = (model_name, revisions, success, error) => (dispatch) => {
-  var data = new FormData()
+// download a TSV from magma via Timur
 
-  for (var record_name in revisions) {
-    var revision = revisions[record_name]
-    for (var attribute_name in revision) {
-      if (Array.isArray(revision[attribute_name])) {
-        revision[attribute_name].forEach(
-          (value) => data.append(
-            revision_name(
-              model_name,record_name,attribute_name
-            )+'[]', value
-          )
-        )
-      }
-      else
-        data.append(
-          revision_name(model_name,record_name,attribute_name),
-          revision[attribute_name]
-        )
-    }
+export const requestTSV = (model_name,filter) =>
+  (dispatch) => {
+    getTSVForm({ model_name, filter, record_names: 'all' })
   }
-
-  postRevisions(data, new Exchange(dispatch, `revisions-${model_name}`))
-    .then((response) => {
-      consumePayload(dispatch,response)
-      dispatch(
-        discardRevision(
-          record_name,
-          model_name
-        )
-      )
-      if (success != undefined) success()
-    })
-    .catch((e) => {
-       e.response.json().then((response) =>
-         dispatch(
-           showMessages([
-`### The change we sought did not occur.
-
-${response.errors.map((error) => `* ${error}`)}`
-           ])
-         )
-       )
-       if (error != undefined) error()
-      })
-  }
-*/
