+++ app/assets/javascripts/components/plots/plotter.jsx
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { selectManifest } from '../../actions/manifest_actions';
-import { requestConsignments } from '../../actions/consignment_actions';
-import { requestManifests, manifestToReqPayload } from '../../actions/manifest_actions';
+import { requestConsignments, requestManifests, manifestToReqPayload } from '../../actions/manifest_actions';
 import { selectConsignment } from '../../selectors/consignment';
 import { saveNewPlot, deletePlot, savePlot, selectPlot, toggleEditing } from '../../actions/plot_actions';
 import { getAllPlots, getSelectedPlot } from '../../selectors/plot';
